requirements.txt
Flask==3.0.0
selenium==4.16.0
Pillow==10.1.0
gunicorn==21.2.0
